+++ src/css/classes.js
@@ -98,34 +98,3 @@ const uniqueClasses = classes
   .join('\n');
 
 console.log(uniqueClasses);
-
-/*
-CSS module thoughts: Atomic css in the component, es6 import, bundler era
-
-- every individual class is exported from its source file, something like:
-
-  ```js
-  // padding.js
-
-  const p0 = { padding: 0}
-  const p1 = { padding: 0.5rem}
-  const p2 = { padding: 1rem}
-  const p3 = { padding: 2rem}
-  const p4 = { padding: 4rem}
-  
-  export { p0, p1, p2, p3, p4 };
-  ```
-  
-  The code example above is not scalable, and isn't a smart use of js!
-  Much refactoring should happen to make this creation of a style palette
-  more automated and dynamic. Something more akin to:
-  
-  > the export default is a function that takes in some value
-  > provided by the developer during component development.
-
-- there can be groups of classes that can be installed together, ie: flexbox
-- the overall css library can be installed via npm
-- the library and/or individual classes and/or groups of classes
-  can be es6 imported
-
-*/